Add dualization and dual noperthedron. Move some more code to cc-lib/geom.
git-svn-id: file:///home/runner/work/tom7misc-sync-action/tom7misc-sync-action/upstream/trunk@7028 c7e11af0-4c7a-11f1-9edd-117cb6f09d55

+++ cc-lib/geom/latlon-tree_test.cc
@@ -33,7 +33,7 @@ static void InsertApp() {
 
   // Find each node exactly once.
   latlontree.App([&points](LatLon ll, int idx) {
-      CHECK(idx >= 0 && idx < points.size());
+      CHECK(idx >= 0 && idx < (int)points.size());
       CHECK(!points[idx].second) << idx;
       points[idx].second = true;
     });
@@ -55,7 +55,7 @@ static void InsertApp() {
         40233600.0);
 
   for (const auto &[ll, idx, dist_] : res) {
-    CHECK(idx >= 0 && idx < points.size());
+    CHECK(idx >= 0 && idx < (int)points.size());
     CHECK(!points[idx].second) << idx;
     points[idx].second = true;
   }
